+++ Mallos.Ai.sln
@@ -26,6 +26,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "dep", "dep", "{7E1DD600-077
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SadConsole.GoRogue", "dep\SadConsole.GoRogueHelpers\src\SadConsole.GoRogue.csproj", "{052D25D4-FE94-4820-9CDC-98969EA8A616}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DialogTreeSample", "samples\DialogTreeSample\DialogTreeSample.csproj", "{D43FC2CA-317B-4F2D-A754-A70CDAFA449A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,6 +54,10 @@ Global
 		{052D25D4-FE94-4820-9CDC-98969EA8A616}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{052D25D4-FE94-4820-9CDC-98969EA8A616}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{052D25D4-FE94-4820-9CDC-98969EA8A616}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D43FC2CA-317B-4F2D-A754-A70CDAFA449A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D43FC2CA-317B-4F2D-A754-A70CDAFA449A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D43FC2CA-317B-4F2D-A754-A70CDAFA449A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D43FC2CA-317B-4F2D-A754-A70CDAFA449A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -62,6 +68,7 @@ Global
 		{6D061BDC-FBDE-4BA4-8297-9FF317B1AD4D} = {32E8C64F-0AEB-4222-BCF6-0DD9E09A2DCE}
 		{3A05163A-22B9-4CC3-8E95-25992346D17E} = {39D7CB70-98D2-4E3C-9D0B-68033E86C586}
 		{052D25D4-FE94-4820-9CDC-98969EA8A616} = {7E1DD600-0778-40F1-91AB-53E5B5173D82}
+		{D43FC2CA-317B-4F2D-A754-A70CDAFA449A} = {32E8C64F-0AEB-4222-BCF6-0DD9E09A2DCE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4D1D27B2-00B1-4652-A93C-8A352929F9AF}
