+++ src/Mallos.Ai/Dialog/DialogTree.cs
@@ -1,34 +1,27 @@
-namespace Mallos.Ai.Dialog
+﻿namespace Mallos.Ai.Dialog
 {
     using System;
     using System.Collections.Generic;
-
-    public interface IDialogTree
-    {
-        void AddProperty(Guid nodeKey, string key, object value);
-
-        Guid AddNode(string text);
-
-        Guid AddChoice(string text);
-
-        void AddLink(Guid node, params Guid[] nodes);
-    }
 
     public class DialogTree : IDialogTree
     {
         private Dictionary<Guid, DialogEntity> nodes = new Dictionary<Guid, DialogEntity>();
         private Dictionary<Guid, List<Guid>> links = new Dictionary<Guid, List<Guid>>();
 
+        public void AddProperty(Guid nodeKey, string key, object value)
+        {
+        }
+
         public Guid AddNode(string text)
         {
-            var key = Guid.New();
+            var key = Guid.NewGuid();
             nodes[key] = new DialogEntity(false, text);
             return key;
         }
 
         public Guid AddChoice(string text)
         {
-            var key = Guid.New();
+            var key = Guid.NewGuid();
             nodes[key] = new DialogEntity(true, text);
             return key;
         }
@@ -37,7 +30,7 @@ namespace Mallos.Ai.Dialog
         {
             if (links.ContainsKey(node))
             {
-                links[node].Append(nodes);
+                links[node].AddRange(nodes);
             }
             else
             {
