+++ src/Mallos.Ai/Dialog/DialogTreeRunner.cs
@@ -1,21 +1,5 @@
-namespace Mallos.Ai.Dialog
+﻿namespace Mallos.Ai.Dialog
 {
-    public interface IDialogTextProcessor
-    {
-        string Process(string text, Blackboard blackboard);
-    }
-
-    public interface IDialogTreeRunner
-    {
-        bool IsActive { get; }
-
-        bool IsChoice { get; }
-
-        string[] CurrentOptions { get; }
-
-        bool Next(int index = 0);
-    }
-
     public class DialogTreeRunner : IDialogTreeRunner
     {
         public bool IsActive { get; private set; }
@@ -28,10 +12,9 @@ namespace Mallos.Ai.Dialog
             DialogTree dialogTree,
             IDialogTextProcessor[] textProcessors = null)
         {
-            
         }
 
-        public bool Act(int index = 0)
+        public bool Next(int index = 0)
         {
             return true;
         }
