+++ src/Mallos.Ai/Dialog/IDialogTextProcessor.cs
@@ -1,0 +1,7 @@
+﻿namespace Mallos.Ai.Dialog
+{
+    public interface IDialogTextProcessor
+    {
+        string Process(string text, Blackboard blackboard);
+    }
+}
